completed user roles

+++ app/controllers/wikis_controller.rb
@@ -5,14 +5,17 @@ class WikisController < ApplicationController
 
   def show
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
   end
 
   def new
     @wiki = Wiki.new
+    authorize @wiki
   end
 
   def create
-    @wiki = Wiki.new(params.require(:wiki).permit(:title, :body, :private))
+    @wiki = Wiki.new(wiki_params)
+    authorize @wiki
 
     if @wiki.save
       flash[:notice] = "Wiki was saved."
@@ -25,11 +28,14 @@ class WikisController < ApplicationController
 
   def edit
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
   end
 
   def update
      @wiki = Wiki.find(params[:id])
-     if @wiki.update_attributes(params.require(:wiki).permit(:title, :body, :private))
+     authorize @wiki
+
+     if @wiki.update_attributes(wiki_params)
        flash[:notice] = "Wiki was updated."
        redirect_to @wiki
      else
@@ -38,16 +44,23 @@ class WikisController < ApplicationController
      end
    end
 
-   def destroy
-     @wiki = Wiki.find(params[:id])
+  def destroy
+    @wiki = Wiki.find(params[:id])
+    authorize @wiki
+    if @wiki.destroy
+      flash[:notice] = "\"#{@wiki.title}\" was deleted successfully."
+      redirect_to wikis_path
+    else
+      flash[:error] = "There was an error deleting the wiki."
+      render :show
+    end
+  end
 
-     if @wiki.destroy
-       flash[:notice] = "\"#{@wiki.title}\" was deleted successfully."
-       redirect_to wikis_path
-     else
-       flash[:error] = "There was an error deleting the wiki."
-       render :show
-     end
-   end
+private
+
+  def wiki_params
+    params.require(:wiki).permit(:title, :body, :private)
+  end
+
 
 end
